Remove database constraint on User fk
The UserOpenID model has a fk to the User model. In a multi-tenant
application that uses PostgreSQL schemas to separate tenant data and
that stores User records in both the public schema and the the tenant
schemas a constraint is placed on this field forcing it to link to the
public schema. This breaks the OpenID authentication for Users in the
tenant schemas. Therefore the contraint has been disabled.